requirements.txt
pyro-ppl
numpy
flask
networkx
